+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added emoji count check to profile data validator.
 - Check mention and link counts when creating post.
+- Disable transaction monitor tasks if blockchain integration is disabled.
 
 ## [1.20.0] - 2023-03-07