+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Check mention and link counts when creating post.
 - Disable transaction monitor tasks if blockchain integration is disabled.
 - Allow multiple configurations in `blockchains` array.
+- Re-fetch object if `attributedTo` value doesn't match `actor` of `Create` activity.
 
 ## [1.20.0] - 2023-03-07