+++ CHANGELOG.md
@@ -21,6 +21,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Re-fetch object if `attributedTo` value doesn't match `actor` of `Create` activity.
 - Added actor validation to `Update(Note)` and `Undo(Follow)` handlers.
 
+### Fixed
+
+- Fixed database query error in `Create` activity handler.
+
 ## [1.20.0] - 2023-03-07
 
 ### Added